+++ shiny/ui/_output.py
@@ -264,7 +264,7 @@ def output_text(
     See Also
     --------
     * :class:`~shiny.render.text`
-    * :func:`~shiny.ui.output_text_verbatim`
+    * :func:`~shiny.ui.output_code`
     """
 
     if not container:
@@ -294,17 +294,10 @@ def output_code(id: str, placeholder: bool = True) -> Tag:
     :
         A UI element
 
-    Note
-    ----
-    This function is currently the same as :func:`~shiny.ui.output_text_verbatim`, but
-    this may change in future versions of Shiny.
-
     See Also
     --------
-    * :class:`~shiny.render.text`
+    * :class:`~shiny.render.code`
     * :func:`~shiny.ui.output_text`
-    * :func:`~shiny.ui.output_code`
-
     """
 
     cls = "shiny-text-output" + (" noplaceholder" if not placeholder else "")
@@ -314,11 +307,12 @@ def output_code(id: str, placeholder: bool = True) -> Tag:
 @add_example(example_name="input_text")
 def output_text_verbatim(id: str, placeholder: bool = False) -> Tag:
     """
-    Deprecated. Create a output container for some text.
+    Deprecated. Use :func:`~shiny.ui.output_code` (for monospaced text) or
+    :func:`~shiny.ui.output_text` (for plain text) instead.
 
-    Place a :class:`~shiny.render.text` result in the user interface.
-    Differs from :func:`~shiny.ui.output_text` in that it wraps the text in a
-    fixed-width container with a gray-ish background color and border.
+    Create a output container for some text. Place a :class:`~shiny.render.text` result
+    in the user interface. Differs from :func:`~shiny.ui.output_text` in that it wraps
+    the text in a fixed-width container with a gray-ish background color and border.
 
     Parameters
     ----------
@@ -336,8 +330,7 @@ def output_text_verbatim(id: str, placeholder: bool = False) -> Tag:
 
     See Also
     --------
-    * :class:`~shiny.render.text`
-    * :class:`~shiny.render.code`
+    * :func:`~shiny.ui.output_code`
     * :func:`~shiny.ui.output_text`
 
     Example
@@ -346,9 +339,10 @@ def output_text_verbatim(id: str, placeholder: bool = False) -> Tag:
     """
 
     warn_deprecated(
-        "`ui.output_text_verbatim()` was deprecated in v1.6.0."
-        "Please use `ui.output_text()` / `@render.text` to create an output container for plain text "
-        "or `ui.output_code()` / `@render.code` to create an output container for monospaced text."
+        "`ui.output_text_verbatim()` was deprecated in v1.8.0. "
+        "Please use `ui.output_code()` / `@render.code` to create an output container "
+        "for monospaced text, or `ui.output_text()` / `@render.text` to create an "
+        "output container for plain text."
     )
 
     cls = "shiny-text-output" + (" noplaceholder" if not placeholder else "")